Added a seed data function for startup. Implemented migrations to create tables in db

+++ Lab3WebMvc/Models/Movie.cs
@@ -9,7 +9,7 @@ namespace Lab3WebMvc.Models
         [Key]
         public int MovieId { get; set; }
         public string Title { get; set;}
-        public float Starting { get; set;}
+        public double Starting { get; set;}
         public virtual IList<Ticket> Tickets { get; set;}
     }
 }
